third_party/WebKit/LayoutTests/fast/encoding/api/legacy-encode.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/shared.js"></script>
<script>

description("Non-UTF encodings supported only for decode, not encode");

encodings_table.forEach(function(section) {
    section.encodings.forEach(function(encoding) {
        if (encoding.name === "replacement")
            return;
        if (utf_encodings.indexOf(encoding.name) !== -1) {
            shouldBeEqualToString("new TextDecoder(" + JSON.stringify(encoding.name) + ").encoding", encoding.name);
            shouldBeEqualToString("new TextEncoder(" + JSON.stringify(encoding.name) + ").encoding", encoding.name);
        } else {
            shouldBeEqualToString("new TextDecoder(" + JSON.stringify(encoding.name) + ").encoding", encoding.name);
            shouldThrow("new TextEncoder(" + JSON.stringify(encoding.name) + ").encoding");
        }
    });
});

</script>
<script src="../../js/resources/js-test-post.js"></script>
