third_party/WebKit/LayoutTests/fast/encoding/api/legacy-encode-expected.txt

Non-UTF encodings supported only for decode, not encode

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new TextDecoder("utf-8").encoding is "utf-8"
PASS new TextEncoder("utf-8").encoding is "utf-8"
FAIL new TextDecoder("ibm866").encoding should be ibm866. Threw exception TypeError: Type error
PASS new TextEncoder("ibm866").encoding threw exception TypeError: Type error.
PASS new TextDecoder("iso-8859-2").encoding is "iso-8859-2"
PASS new TextEncoder("iso-8859-2").encoding threw exception TypeError: Type error.
PASS new TextDecoder("iso-8859-3").encoding is "iso-8859-3"
PASS new TextEncoder("iso-8859-3").encoding threw exception TypeError: Type error.
PASS new TextDecoder("iso-8859-4").encoding is "iso-8859-4"
PASS new TextEncoder("iso-8859-4").encoding threw exception TypeError: Type error.
PASS new TextDecoder("iso-8859-5").encoding is "iso-8859-5"
PASS new TextEncoder("iso-8859-5").encoding threw exception TypeError: Type error.
PASS new TextDecoder("iso-8859-6").encoding is "iso-8859-6"
PASS new TextEncoder("iso-8859-6").encoding threw exception TypeError: Type error.
PASS new TextDecoder("iso-8859-7").encoding is "iso-8859-7"
PASS new TextEncoder("iso-8859-7").encoding threw exception TypeError: Type error.
PASS new TextDecoder("iso-8859-8").encoding is "iso-8859-8"
PASS new TextEncoder("iso-8859-8").encoding threw exception TypeError: Type error.
PASS new TextDecoder("iso-8859-8-i").encoding is "iso-8859-8-i"
PASS new TextEncoder("iso-8859-8-i").encoding threw exception TypeError: Type error.
PASS new TextDecoder("iso-8859-10").encoding is "iso-8859-10"
PASS new TextEncoder("iso-8859-10").encoding threw exception TypeError: Type error.
PASS new TextDecoder("iso-8859-13").encoding is "iso-8859-13"
PASS new TextEncoder("iso-8859-13").encoding threw exception TypeError: Type error.
PASS new TextDecoder("iso-8859-14").encoding is "iso-8859-14"
PASS new TextEncoder("iso-8859-14").encoding threw exception TypeError: Type error.
PASS new TextDecoder("iso-8859-15").encoding is "iso-8859-15"
PASS new TextEncoder("iso-8859-15").encoding threw exception TypeError: Type error.
PASS new TextDecoder("iso-8859-16").encoding is "iso-8859-16"
PASS new TextEncoder("iso-8859-16").encoding threw exception TypeError: Type error.
PASS new TextDecoder("koi8-r").encoding is "koi8-r"
PASS new TextEncoder("koi8-r").encoding threw exception TypeError: Type error.
PASS new TextDecoder("koi8-u").encoding is "koi8-u"
PASS new TextEncoder("koi8-u").encoding threw exception TypeError: Type error.
PASS new TextDecoder("macintosh").encoding is "macintosh"
PASS new TextEncoder("macintosh").encoding threw exception TypeError: Type error.
PASS new TextDecoder("windows-874").encoding is "windows-874"
PASS new TextEncoder("windows-874").encoding threw exception TypeError: Type error.
PASS new TextDecoder("windows-1250").encoding is "windows-1250"
PASS new TextEncoder("windows-1250").encoding threw exception TypeError: Type error.
PASS new TextDecoder("windows-1251").encoding is "windows-1251"
PASS new TextEncoder("windows-1251").encoding threw exception TypeError: Type error.
PASS new TextDecoder("windows-1252").encoding is "windows-1252"
PASS new TextEncoder("windows-1252").encoding threw exception TypeError: Type error.
PASS new TextDecoder("windows-1253").encoding is "windows-1253"
PASS new TextEncoder("windows-1253").encoding threw exception TypeError: Type error.
PASS new TextDecoder("windows-1254").encoding is "windows-1254"
PASS new TextEncoder("windows-1254").encoding threw exception TypeError: Type error.
PASS new TextDecoder("windows-1255").encoding is "windows-1255"
PASS new TextEncoder("windows-1255").encoding threw exception TypeError: Type error.
PASS new TextDecoder("windows-1256").encoding is "windows-1256"
PASS new TextEncoder("windows-1256").encoding threw exception TypeError: Type error.
PASS new TextDecoder("windows-1257").encoding is "windows-1257"
PASS new TextEncoder("windows-1257").encoding threw exception TypeError: Type error.
PASS new TextDecoder("windows-1258").encoding is "windows-1258"
PASS new TextEncoder("windows-1258").encoding threw exception TypeError: Type error.
PASS new TextDecoder("x-mac-cyrillic").encoding is "x-mac-cyrillic"
PASS new TextEncoder("x-mac-cyrillic").encoding threw exception TypeError: Type error.
PASS new TextDecoder("gbk").encoding is "gbk"
PASS new TextEncoder("gbk").encoding threw exception TypeError: Type error.
PASS new TextDecoder("gb18030").encoding is "gb18030"
PASS new TextEncoder("gb18030").encoding threw exception TypeError: Type error.
PASS new TextDecoder("hz-gb-2312").encoding is "hz-gb-2312"
PASS new TextEncoder("hz-gb-2312").encoding threw exception TypeError: Type error.
PASS new TextDecoder("big5").encoding is "big5"
PASS new TextEncoder("big5").encoding threw exception TypeError: Type error.
PASS new TextDecoder("euc-jp").encoding is "euc-jp"
PASS new TextEncoder("euc-jp").encoding threw exception TypeError: Type error.
PASS new TextDecoder("iso-2022-jp").encoding is "iso-2022-jp"
PASS new TextEncoder("iso-2022-jp").encoding threw exception TypeError: Type error.
PASS new TextDecoder("shift_jis").encoding is "shift_jis"
PASS new TextEncoder("shift_jis").encoding threw exception TypeError: Type error.
FAIL new TextDecoder("euc-kr").encoding should be euc-kr. Was windows-949.
PASS new TextEncoder("euc-kr").encoding threw exception TypeError: Type error.
PASS new TextDecoder("iso-2022-kr").encoding is "iso-2022-kr"
PASS new TextEncoder("iso-2022-kr").encoding threw exception TypeError: Type error.
PASS new TextDecoder("utf-16be").encoding is "utf-16be"
PASS new TextEncoder("utf-16be").encoding is "utf-16be"
PASS new TextDecoder("utf-16le").encoding is "utf-16le"
PASS new TextEncoder("utf-16le").encoding is "utf-16le"
PASS new TextDecoder("x-user-defined").encoding is "x-user-defined"
PASS new TextEncoder("x-user-defined").encoding threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
